#632 - Move project container into project page folder
