+++ src/components/projects/wbs-details/project-page/project-container/project-details/project-details.module.css
