+++ src/components/projects/wbs-details/project-page/project-container/rules-list/rules-list.test.tsx
@@ -4,7 +4,7 @@
  */
 
 import { render, screen } from '@testing-library/react';
-import { exampleProject1 } from '../../../../../test-support/test-data/projects.stub';
+import { exampleProject1 } from '../../../../../../test-support/test-data/projects.stub';
 import RulesList from './rules-list';
 
 describe('Rendering Work Package Rules Component', () => {
