+++ src/components/projects/wbs-details/project-page/project-container/work-package-summary/work-package-summary.test.tsx
@@ -4,13 +4,13 @@
  */
 
 import { WorkPackage } from 'utils';
-import { render, screen, routerWrapperBuilder } from '../../../../../test-support/test-utils';
-import { wbsPipe, listPipe, endDatePipe, datePipe } from '../../../../../shared/pipes';
+import { render, screen, routerWrapperBuilder } from '../../../../../../test-support/test-utils';
+import { wbsPipe, listPipe, endDatePipe, datePipe } from '../../../../../../shared/pipes';
 import {
   exampleWorkPackage1,
   exampleWorkPackage2,
   exampleWorkPackage3
-} from '../../../../../test-support/test-data/work-packages.stub';
+} from '../../../../../../test-support/test-data/work-packages.stub';
 import WorkPackageSummary from './work-package-summary';
 
 // Sets up the component under test with the desired values and renders it
